# 8305774: String.join(CharSequence, Iterable) can be optimized if Iterable is a Collection