--- a/src/java.base/share/classes/java/lang/String.java
+++ b/src/java.base/share/classes/java/lang/String.java
@@ -36,7 +36,6 @@ import java.nio.CharBuffer;
 import java.nio.charset.*;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Comparator;
 import java.util.Formatter;
 import java.util.List;
@@ -3455,7 +3454,14 @@ public final class String
         Objects.requireNonNull(delimiter);
         Objects.requireNonNull(elements);
         var delim = delimiter.toString();
-        var elems = new String[elements instanceof Collection<?> c ? c.size() : 8];
+        String[] elems;
+        long initLen = elements.spliterator().estimateSize();
+        if (initLen <= 8L || initLen >= ArraysSupport.SOFT_MAX_ARRAY_LENGTH) {
+            // This usually means the size of the Iterable cannot be computed easily.
+            elems = new String[8];
+        } else {
+            elems = new String[(int)initLen];
+        }
         int size = 0;
         for (CharSequence cs: elements) {
             if (size >= elems.length) {