@JimLaskey 2023-04-08
It would be good to show the performance improvement via a few jmh runs.