@liach 2023-04-08
Have you considered the performance impact on concurrent collections, that their size() computation is non-constant, and the array still may need resizing later (which you've kept, good)