@yuantj 2023-04-08
@liach Thanks for reminding that. Some collections such as `java.util.concurrent.ConcurrentLinkedDeque` requires O(n) time to compute the `size`. I will benchmark it as well.