@jmehrens 2023-04-09
What about using Iterable::spliterator instead of Iterable::iterator (enhanced for loop)?  Spliterator would allow you to call Spliterator::estimateSize and then to default to 8 if less than 8, cast and use if less than or equal to ArraysSupport.SOFT_MAX_ARRAY_LENGTH, or throw OutOfMemoryError if greater than Integer.MAX_VALUE.  The array growth code would have to use ArraysSupport.newLength to deal with huge sizes.

I'm not sure if bootstrapping issues will prevent you from using lambdas expressions to traverse the Spliterator.  If lambdas expressions are forbidden then this approach will introduce some code smell by having to use inner classes.