@yuantj 2023-04-09
I've just added the benchmark code, and here is the result tested on my computer.

Before change:

```plain
Benchmark                                           Mode  Cnt          Score          Error  Units
StringJoinWithIterable.joinWithArrayList            avgt   30  146820983.745 ±  6654877.812  ns/op
StringJoinWithIterable.joinWithIterable             avgt   30  145236859.688 ±  5683967.478  ns/op
StringJoinWithIterable.joinWithLinkedBlockingDeque  avgt   30  250139433.967 ± 14381158.389  ns/op
StringJoinWithIterable.joinWithLinkedHashSet        avgt   30  167718287.313 ±  7394905.875  ns/op
StringJoinWithIterable.joinWithLinkedTransferQueue  avgt   30  153245459.832 ±  7972716.070  ns/op
```

After change:

```plain
Benchmark                                           Mode  Cnt          Score          Error  Units
StringJoinWithIterable.joinWithArrayList            avgt   30  138343729.406 ±  5959140.498  ns/op
StringJoinWithIterable.joinWithIterable             avgt   30  149566458.195 ±  9755905.704  ns/op
StringJoinWithIterable.joinWithLinkedBlockingDeque  avgt   30  247384811.927 ± 14647765.134  ns/op
StringJoinWithIterable.joinWithLinkedHashSet        avgt   30  153244492.767 ±  4589884.301  ns/op
StringJoinWithIterable.joinWithLinkedTransferQueue  avgt   30  157162685.594 ±  4330658.539  ns/op

```

It seems that the improvement is not notable :-)