@ExE-Boss 2023-04-09
You can also use [`Spliterators.iterator(Spliterator)`] to convert the [`Spliterator`] into an [`Iterator`] after the size computation without needing to change the loop.

[`Spliterators.iterator(Spliterator)`]: https://docs.oracle.com/en/java/javase/20/docs/api/java.base/java/util/Spliterators.html#iterator%28java.util.Spliterator%29
[`Spliterator`]: https://docs.oracle.com/en/java/javase/20/docs/api/java.base/java/util/Spliterator.html
[`Iterator`]: https://docs.oracle.com/en/java/javase/20/docs/api/java.base/java/util/Iterator.html