@yuantj 2023-04-10
> @yuantj Alternatively, you can probably try working on the `toArray` result of a collection than to allocate a new String array. This might be more efficient for some implementations like ArrayList as well, in addition to benefiting concurrent collections.

The method accepts collections of nullable CharSequences, which may not be Strings.