@yuantj 2023-04-10
@liach  Did you mean this?

```Java
public static String join(CharSequence delimiter,
            Iterable<? extends CharSequence> elements) {
    if (elements instanceof Collection<?> c) {
        try {
            return String.join(delimiter, c.toArray(new CharSequence[0]));
        } catch (ArrayStoreException e) {
            throw new ClassCastException(e.getMessage());
        }
    } else {
        // old implementations for non-collection iterables.
    }
}
```