@yuantj 2023-04-10
What about setting the initial length to this?

```Java
String[] elems;
long initLen = elements.spliterator().estimateSize();
if (initLen < 8L || initLen > Integer.MAX_VALUE - 8L) {
    // This usually means the size of the Iterable cannot be computed easily.
    elems = new String[8];
} else {
    elems = new String[(int)initLen];
}
```

This doesn't require that `elements` is an instance of `Collection`.