@liach 2023-04-10
Nah. I mean like:
```java
    public static String join(CharSequence delimiter,
            Iterable<? extends CharSequence> elements) {
        Objects.requireNonNull(delimiter);
        Objects.requireNonNull(elements);
        var delim = delimiter.toString();
        Object[] elems;
        final int size;
        if (elements instanceof Collection<?> c) {
            elems = c.toArray();
            size = elems.length;
            for (int i = 0; i < size; i++) {
                elems[i] = String.valueOf(elems[i]);
            }
        } else {
            elems = new String[elements instanceof Collection<?> c ? c.size() : 8]; // or whatever spliterator you decide to use
            size = 0;
            for (CharSequence cs: elements) {
                if (size >= elems.length) {
                    elems = Arrays.copyOf(elems, elems.length << 1);
                }
                elems[size++] = String.valueOf(cs);
            }
        }
        return join("", "", delim, elems, size);
    }
// ...
static String join(String prefix, String suffix, String delimiter, Object[] elements, int size) { // change array type to Object[], cast on access elements
```